src/requirements.txt
pydantic-settings==2.1.0
python-dotenv==1.0.0
psycopg2-binary==2.9.9
requests==2.31.0
python-json-logger==2.0.7